🗁 When adding new file, set as startup file
We previously generated one entry in launchSettings.json for each top-level
file. This was problematic because now we had to find a way to automatically
select the right entry whenever the active document changed. This was not
possible using VS extensibility APIs as-is, so a better and simpler approach
is to just generate ONE entry in the file, which matches the current active
document. Also, we update the .user with the ActiveDebugProfile at the same
time. This avoids some timing issues we had in the past, and simplifies further.

This means we now don't really need source generator capability anymore,
since we do everything from MSBuild now. We'll still check for the right C#
version since otherwise folks could install the package on lower versions of
VS and expect it to work (which it wouldn't).

Fixes #9.
