--- src/SmallSharp.Build/SmallSharp.Build.csproj
@@ -1,25 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
-    <PackFolder>build\netstandard2.0</PackFolder>
-    <PackNone>true</PackNone>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" Pack="false" />
-    <PackageReference Include="NuGetizer" Version="0.6.0" />
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.Build.Tasks.Core" Version="16.8.0" Pack="false" />
-    <PackageReference Include="EnvDTE" Version="16.7.30508.193" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.VisualStudio.OLE.Interop" Version="16.7.30328.74" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.VisualStudio.Shell.Interop" Version="16.7.30328.74" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.VisualStudio.Shell.Interop.12.0" Version="16.7.30328.74" PrivateAssets="all" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" PrivateAssets="all" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="SmallSharp.targets" CopyToOutputDirectory="PreserveNewest" />
-  </ItemGroup>
-
-</Project>
