+++ src/SmallSharp/DisposableAction.cs
