+++ src/SmallSharp/SmallSharp.csproj
@@ -1,26 +1,32 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>net472</TargetFramework>
     <PackageId>SmallSharp</PackageId>
     <Description>Create, edit and run multiple C# 9.0 top-level programs in the same project 😍</Description>
-    <PackFolder>analyzers\dotnet\cs</PackFolder>
+    <PackFolder>build\netstandard2.0</PackFolder>
+    <PackNone>true</PackNone>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" Pack="false" />
     <PackageReference Include="NuGetizer" Version="0.6.0" />
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.8.0" Pack="false" />
+    <PackageReference Include="Microsoft.Build.Tasks.Core" Version="16.8.0" Pack="false" />
+    <PackageReference Include="EnvDTE" Version="16.7.30508.193" PrivateAssets="all" />
+    <PackageReference Include="Microsoft.VisualStudio.OLE.Interop" Version="16.7.30328.74" PrivateAssets="all" />
+    <PackageReference Include="Microsoft.VisualStudio.Shell.Interop" Version="16.7.30328.74" PrivateAssets="all" />
+    <PackageReference Include="Microsoft.VisualStudio.Shell.Interop.12.0" Version="16.7.30328.74" PrivateAssets="all" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" PrivateAssets="all" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SmallSharp.Build\SmallSharp.Build.csproj" ReferenceOutputAssembly="false" />
+    <UpToDateCheckInput Include="SmallSharp.targets" />
+    <None Include="..\_._" PackFolder="lib\netstandard2.0" Visible="false" />
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="@(None)" PackFolder="build\netstandard2.0" />
-    <None Include="..\_._" PackFolder="lib\netstandard2.0" />
+    <None Update="SmallSharp.targets" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
 
 </Project>
